Type cast is redundant

+++ src/PublicSuffixList.php
@@ -226,7 +226,7 @@ class PublicSuffixList
         $cacheFile = $this->getCacheFileName($url);
         if (file_exists($cacheFile)) {
             $cachedTree = file_get_contents($cacheFile);
-            if ((int) PHP_VERSION_ID < 70000) {
+            if (PHP_VERSION_ID < 70000) {
                 return unserialize($cachedTree);
             }
             return unserialize($cachedTree, array('allowed_classes' => false));
